fix: fix delete functions

+++ src/service/brandService.js
@@ -1,4 +1,5 @@
 const {Brand} = require("../models/models");
+const ApiError = require("../error/ApiError");
 
 class BrandService {
   async create(name) {
@@ -9,6 +10,11 @@ class BrandService {
   }
 
   async delete(name) {
+    const brand = await Brand.findOne({where: { name }});
+    if (!brand) {
+      throw ApiError.badRequest(`Бренд с именем ${name} не найден!`)
+    }
+
     await Brand.destroy({where: { name }});
   }
 }
