+++ src/service/typeService.js
@@ -1,4 +1,5 @@
 const {Type} = require('../models/models');
+const ApiError = require("../error/ApiError");
 
 class TypeService {
   async create(name) {
@@ -6,6 +7,11 @@ class TypeService {
   }
 
   async delete(name) {
+    const type = await Type.findOne({where: { name }});
+    if (!type) {
+      throw ApiError.badRequest(`Тип с названием ${name} не найден`);
+    }
+
     return await Type.destroy({where: { name }});
   }
